NSE_Data/nse-azure-function/requirements.txt
azure-functions
azure-storage-blob
requests
pandas
python-dateutil
python-dotenv
playwright
chromium